requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
bcrypt==4.0.1
Pillow==10.0.0
python-dotenv==1.0.0
pytz==2023.3